+++ Runtime/ComponentHeader/HeaderDrawAttribute.cs
@@ -1,10 +1,14 @@
 using System;
 using System.Diagnostics;
+#if SAINTSFIELD_IMPLICIT_USE
 using JetBrains.Annotations;
+#endif
 
 namespace SaintsField.ComponentHeader
 {
-    [MeansImplicitUse]
+#if SAINTSFIELD_IMPLICIT_USE
+    [MeansImplicitUse]  // https://github.com/TylerTemp/SaintsField/pull/171#issuecomment-3680042013
+#endif
     [Conditional("UNITY_EDITOR")]
     [AttributeUsage(AttributeTargets.Method)]
     public class HeaderDrawAttribute: AbsComponentHeaderAttribute