+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Fix: `ShowInInspector` inside `struct` can not read a real-time but cached value
 
 ## 5.7.4 ##
 
-1.  Change: `Button` no longer use `MeansImplicitUse`. Add `SAINTSFIELD_BUTTON_IMPLICIT_USE` macro instead.
+1.  Change: `Button` no longer use `MeansImplicitUse`. Add `SAINTSFIELD_IMPLICIT_USE` macro instead.
 2.  Fix: `OnValueChanged`, `Button` (including inline buttons) now can in-place change a struct's value [#240](https://github.com/TylerTemp/SaintsField/issues/240), [#287](https://github.com/TylerTemp/SaintsField/issues/287)
 
 ## 5.7.3 ##